Moved add mode to unitary function into matrix utils file - Also reworked all imports to remove the imports from the sdk init. Now components should be imported from their respective sub-components within the sdk module.
