+++ lightworks/emulator/simulation/quick_sampler.py
@@ -15,7 +15,8 @@
 from ..backend import Backend
 from ..utils import fock_basis, ModeMismatchError
 from ..results import SamplingResult
-from ...sdk import State, Circuit
+from ...sdk.state import State
+from ...sdk.circuit import Circuit
 
 import numpy as np
 from random import random
