+++ lightworks/sdk/__init__.py
@@ -11,9 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from .circuit import *
-from .visualisation import *
-from .state import *
-from .utils import *
-from .optimisation import *
